+++ Sources/Analytics/UserInterface/ContainerPageViewTracking.swift
@@ -9,3 +9,10 @@ import UIKit
 public protocol ContainerPageViewTracking {
     var activeChildren: [UIViewController] { get }
 }
+
+extension UINavigationController: ContainerPageViewTracking {
+    public var activeChildren: [UIViewController] {
+        guard let topViewController else { return [] }
+        return [topViewController]
+    }
+}
